refactor: visualisation tests
