+++ engine/runtime/src/test/scala/org/enso/interpreter/test/instrument/RuntimeServerTest.scala
@@ -10,7 +10,6 @@ import org.enso.polyglot.data.TypeGraph
 import org.enso.polyglot.runtime.Runtime.Api
 import org.enso.text.editing.model
 import org.enso.text.editing.model.TextEdit
-import org.enso.text.{ContentVersion, Sha3_224VersionCalculator}
 import org.graalvm.polyglot.Context
 import org.graalvm.polyglot.io.MessageEndpoint
 import org.scalatest.BeforeAndAfterEach
@@ -321,9 +320,6 @@ class RuntimeServerTest
     }
 
   }
-
-  def contentsVersion(content: String): ContentVersion =
-    Sha3_224VersionCalculator.evalVersion(content)
 
   override protected def beforeEach(): Unit = {
     context = new TestContext("Test")
@@ -3005,1095 +3001,6 @@ class RuntimeServerTest
       context.executionComplete(contextId)
     )
     context.consumeOut shouldEqual List()
-  }
-
-  it should "emit visualisation update when expression is computed" in {
-    val idMain     = context.Main.metadata.addItem(78, 1)
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(6) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      TestMessages.update(contextId, idMain, Constants.INTEGER),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          idMain,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-    val attachVisualisationResponses = context.receive(3)
-    attachVisualisationResponses should contain allOf (
-      Api.Response(requestId, Api.VisualisationAttached()),
-      context.executionComplete(contextId)
-    )
-    val Some(data) = attachVisualisationResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `idMain`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data.sameElements("50".getBytes) shouldBe true
-
-    // recompute
-    context.send(
-      Api.Request(requestId, Api.RecomputeContextRequest(contextId, None))
-    )
-
-    val recomputeResponses = context.receive(3)
-    recomputeResponses should contain allOf (
-      Api.Response(requestId, Api.RecomputeContextResponse(contextId)),
-      context.executionComplete(contextId)
-    )
-    val Some(data2) = recomputeResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `idMain`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data2.sameElements("50".getBytes) shouldBe true
-  }
-
-  it should "emit visualisation update when expression is cached" in {
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(5) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          context.Main.idMainX,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-    val attachVisualisationResponses = context.receive(2)
-    attachVisualisationResponses should contain(
-      Api.Response(requestId, Api.VisualisationAttached())
-    )
-    val expectedExpressionId = context.Main.idMainX
-    val Some(data) = attachVisualisationResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data.sameElements("6".getBytes) shouldBe true
-
-    // recompute
-    context.send(
-      Api.Request(requestId, Api.RecomputeContextRequest(contextId, None))
-    )
-    context.receive(2) should contain allOf (
-      Api.Response(requestId, Api.RecomputeContextResponse(contextId)),
-      context.executionComplete(contextId)
-    )
-
-    // recompute invalidating x
-    context.send(
-      Api.Request(
-        requestId,
-        Api.RecomputeContextRequest(
-          contextId,
-          Some(
-            Api.InvalidatedExpressions.Expressions(Vector(context.Main.idMainX))
-          )
-        )
-      )
-    )
-    val recomputeResponses2 = context.receive(3)
-    recomputeResponses2 should contain allOf (
-      Api.Response(requestId, Api.RecomputeContextResponse(contextId)),
-      context.executionComplete(contextId)
-    )
-    val Some(data2) = recomputeResponses2.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data2.sameElements("6".getBytes) shouldBe true
-  }
-
-  it should "emit visualisation update without value update" in {
-    val contents   = context.Main.code
-    val moduleName = "Test.Main"
-    val mainFile   = context.writeMain(contents)
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // open files
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-    context.receiveNone shouldEqual None
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-
-    context.receive(5) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualization
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          context.Main.idMainX,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-    val attachVisualisationResponses = context.receive(2)
-    attachVisualisationResponses should contain(
-      Api.Response(requestId, Api.VisualisationAttached())
-    )
-    val expectedExpressionId = context.Main.idMainX
-    val Some(data) = attachVisualisationResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data.sameElements("6".getBytes) shouldBe true
-
-    // Modify the file
-    context.send(
-      Api.Request(
-        Api.EditFileNotification(
-          mainFile,
-          Seq(
-            TextEdit(
-              model.Range(model.Position(4, 8), model.Position(4, 9)),
-              "5"
-            )
-          )
-        )
-      )
-    )
-
-    val editFileResponse = context.receive(1)
-    val Some(data1) = editFileResponse.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data1.sameElements("5".getBytes) shouldBe true
-  }
-
-  it should "be able to modify visualisations" in {
-    val contents = context.Main.code
-    val mainFile = context.writeMain(contents)
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    // open files
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-    context.receiveNone shouldEqual None
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer("Test.Main", "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(5) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          context.Main.idMainX,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-
-    val attachVisualisationResponses = context.receive(2)
-    attachVisualisationResponses should contain(
-      Api.Response(requestId, Api.VisualisationAttached())
-    )
-    val expectedExpressionId = context.Main.idMainX
-    val Some(data) = attachVisualisationResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data.sameElements("6".getBytes) shouldBe true
-
-    // modify visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.ModifyVisualisation(
-          visualisationId,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.incAndEncode x"
-          )
-        )
-      )
-    )
-    val modifyVisualisationResponses = context.receive(2)
-    modifyVisualisationResponses should contain(
-      Api.Response(requestId, Api.VisualisationModified())
-    )
-    val Some(dataAfterModification) =
-      modifyVisualisationResponses.collectFirst {
-        case Api.Response(
-              None,
-              Api.VisualisationUpdate(
-                Api.VisualisationContext(
-                  `visualisationId`,
-                  `contextId`,
-                  `expectedExpressionId`
-                ),
-                data
-              )
-            ) =>
-          data
-      }
-    dataAfterModification.sameElements("7".getBytes) shouldBe true
-  }
-
-  it should "not emit visualisation updates when visualisation is detached" in {
-    val contents = context.Main.code
-    val mainFile = context.writeMain(contents)
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    // open files
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          context.Main.idMainX,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-    context.receive(3) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.VisualisationAttached()),
-      Api.Response(
-        Api.ExecutionFailed(
-          contextId,
-          Api.ExecutionResult.Failure("Execution stack is empty.", None)
-        )
-      ),
-      context.executionComplete(contextId)
-    )
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer("Test.Main", "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    val pushResponses = context.receive(6)
-    pushResponses should contain allOf (
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      context.executionComplete(contextId)
-    )
-    val expectedExpressionId = context.Main.idMainX
-    val Some(data) =
-      pushResponses.collectFirst {
-        case Api.Response(
-              None,
-              Api.VisualisationUpdate(
-                Api.VisualisationContext(
-                  `visualisationId`,
-                  `contextId`,
-                  `expectedExpressionId`
-                ),
-                data
-              )
-            ) =>
-          data
-      }
-    data.sameElements("6".getBytes) shouldBe true
-
-    // detach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.DetachVisualisation(
-          contextId,
-          visualisationId,
-          context.Main.idMainX
-        )
-      )
-    )
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.VisualisationDetached())
-    )
-
-    // recompute
-    context.send(
-      Api.Request(requestId, Api.RecomputeContextRequest(contextId, None))
-    )
-    context.receive(2) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.RecomputeContextResponse(contextId)),
-      context.executionComplete(contextId)
-    )
-
-    // recompute invalidating x
-    context.send(
-      Api.Request(
-        requestId,
-        Api.RecomputeContextRequest(
-          contextId,
-          Some(
-            Api.InvalidatedExpressions.Expressions(Vector(context.Main.idMainX))
-          )
-        )
-      )
-    )
-    context.receive(2) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.RecomputeContextResponse(contextId)),
-      context.executionComplete(contextId)
-    )
-  }
-
-  it should "not reorder visualization commands" in {
-    val contents = context.Main.code
-    val mainFile = context.writeMain(contents)
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", context.Visualisation.code)
-
-    // open files
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          context.Visualisation.code,
-          true
-        )
-      )
-    )
-    context.receiveNone shouldEqual None
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer("Test.Main", "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(5) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          context.Main.idMainX,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.encode x"
-          )
-        )
-      )
-    )
-
-    val attachVisualisationResponses = context.receive(2)
-    attachVisualisationResponses should contain(
-      Api.Response(requestId, Api.VisualisationAttached())
-    )
-    val expectedExpressionId = context.Main.idMainX
-    val Some(data) = attachVisualisationResponses.collectFirst {
-      case Api.Response(
-            None,
-            Api.VisualisationUpdate(
-              Api.VisualisationContext(
-                `visualisationId`,
-                `contextId`,
-                `expectedExpressionId`
-              ),
-              data
-            )
-          ) =>
-        data
-    }
-    data.sameElements("6".getBytes) shouldBe true
-
-    // modify visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.ModifyVisualisation(
-          visualisationId,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "x -> here.incAndEncode x"
-          )
-        )
-      )
-    )
-    // detach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.DetachVisualisation(
-          contextId,
-          visualisationId,
-          context.Main.idMainX
-        )
-      )
-    )
-    val modifyVisualisationResponses = context.receive(3)
-    modifyVisualisationResponses should contain allOf (
-      Api.Response(requestId, Api.VisualisationModified()),
-      Api.Response(requestId, Api.VisualisationDetached())
-    )
-    val Some(dataAfterModification) =
-      modifyVisualisationResponses.collectFirst {
-        case Api.Response(
-              None,
-              Api.VisualisationUpdate(
-                Api.VisualisationContext(
-                  `visualisationId`,
-                  `contextId`,
-                  `expectedExpressionId`
-                ),
-                data
-              )
-            ) =>
-          data
-      }
-    dataAfterModification.sameElements("7".getBytes) shouldBe true
-  }
-
-  it should "return ModuleNotFound error when attaching visualisation" in {
-    val idMain     = context.Main.metadata.addItem(78, 1)
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(6) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      TestMessages.update(contextId, idMain, Constants.INTEGER),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          idMain,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Undefined",
-            "x -> x"
-          )
-        )
-      )
-    )
-    context.receive(1) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.ModuleNotFound("Test.Undefined"))
-    )
-  }
-
-  it should "return VisualisationExpressionFailed error when attaching visualisation" in {
-    val idMain     = context.Main.metadata.addItem(78, 1)
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(6) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      TestMessages.update(contextId, idMain, Constants.INTEGER),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          idMain,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Main",
-            "here.does_not_exist"
-          )
-        )
-      )
-    )
-    context.receive(1) should contain theSameElementsAs Seq(
-      Api.Response(
-        requestId,
-        Api.VisualisationExpressionFailed(
-          "Method `does_not_exist` of Main could not be found.",
-          Some(
-            Api.ExecutionResult.Diagnostic.error(
-              message = "Method `does_not_exist` of Main could not be found.",
-              stack = Vector(
-                Api.StackTraceElement("<eval>", None, None, None),
-                Api.StackTraceElement("Debug.eval", None, None, None)
-              )
-            )
-          )
-        )
-      )
-    )
-  }
-
-  it should "return visualisation evaluation errors with diagnostic info" in {
-    val idMain     = context.Main.metadata.addItem(78, 1)
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(6) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      TestMessages.update(contextId, idMain, Constants.INTEGER),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          idMain,
-          Api.VisualisationConfiguration(
-            contextId,
-            moduleName,
-            "x -> x.visualise_me"
-          )
-        )
-      )
-    )
-    context.receive(3) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.VisualisationAttached()),
-      Api.Response(
-        Api.VisualisationEvaluationFailed(
-          contextId,
-          visualisationId,
-          idMain,
-          "Method `visualise_me` of 50 (Integer) could not be found.",
-          Some(
-            Api.ExecutionResult.Diagnostic.error(
-              "Method `visualise_me` of 50 (Integer) could not be found.",
-              None,
-              Some(model.Range(model.Position(0, 5), model.Position(0, 19))),
-              None,
-              Vector(
-                Api.StackTraceElement(
-                  "<anonymous>",
-                  None,
-                  Some(
-                    model.Range(model.Position(0, 5), model.Position(0, 19))
-                  ),
-                  None
-                )
-              )
-            )
-          )
-        )
-      ),
-      context.executionComplete(contextId)
-    )
-  }
-
-  it should "return visualisation error with a stack trace" in {
-    val idMain     = context.Main.metadata.addItem(78, 1)
-    val contents   = context.Main.code
-    val mainFile   = context.writeMain(context.Main.code)
-    val moduleName = "Test.Main"
-    val visualisationCode =
-      """
-        |encode x = x.visualise_me
-        |
-        |inc_and_encode x = here.encode x+1
-        |""".stripMargin.linesIterator.mkString("\n")
-
-    val visualisationFile =
-      context.writeInSrcDir("Visualisation", visualisationCode)
-
-    context.send(
-      Api.Request(
-        Api.OpenFileNotification(
-          visualisationFile,
-          visualisationCode,
-          true
-        )
-      )
-    )
-
-    val contextId       = UUID.randomUUID()
-    val requestId       = UUID.randomUUID()
-    val visualisationId = UUID.randomUUID()
-
-    // create context
-    context.send(Api.Request(requestId, Api.CreateContextRequest(contextId)))
-    context.receive shouldEqual Some(
-      Api.Response(requestId, Api.CreateContextResponse(contextId))
-    )
-
-    // Open the new file
-    context.send(
-      Api.Request(Api.OpenFileNotification(mainFile, contents, true))
-    )
-    context.receiveNone shouldEqual None
-
-    // push main
-    val item1 = Api.StackItem.ExplicitCall(
-      Api.MethodPointer(moduleName, "Test.Main", "main"),
-      None,
-      Vector()
-    )
-    context.send(
-      Api.Request(requestId, Api.PushContextRequest(contextId, item1))
-    )
-    context.receive(6) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.PushContextResponse(contextId)),
-      context.Main.Update.mainX(contextId),
-      context.Main.Update.mainY(contextId),
-      context.Main.Update.mainZ(contextId),
-      TestMessages.update(contextId, idMain, Constants.INTEGER),
-      context.executionComplete(contextId)
-    )
-
-    // attach visualisation
-    context.send(
-      Api.Request(
-        requestId,
-        Api.AttachVisualisation(
-          visualisationId,
-          idMain,
-          Api.VisualisationConfiguration(
-            contextId,
-            "Test.Visualisation",
-            "here.inc_and_encode"
-          )
-        )
-      )
-    )
-    context.receive(3) should contain theSameElementsAs Seq(
-      Api.Response(requestId, Api.VisualisationAttached()),
-      Api.Response(
-        Api.VisualisationEvaluationFailed(
-          contextId,
-          visualisationId,
-          idMain,
-          "Method `visualise_me` of 51 (Integer) could not be found.",
-          Some(
-            Api.ExecutionResult.Diagnostic.error(
-              "Method `visualise_me` of 51 (Integer) could not be found.",
-              Some(visualisationFile),
-              Some(model.Range(model.Position(1, 11), model.Position(1, 25))),
-              None,
-              Vector(
-                Api.StackTraceElement(
-                  "Visualisation.encode",
-                  Some(visualisationFile),
-                  Some(
-                    model.Range(model.Position(1, 11), model.Position(1, 25))
-                  ),
-                  None
-                ),
-                Api.StackTraceElement(
-                  "Visualisation.inc_and_encode",
-                  Some(visualisationFile),
-                  Some(
-                    model.Range(model.Position(3, 19), model.Position(3, 34))
-                  ),
-                  None
-                )
-              )
-            )
-          )
-        )
-      ),
-      context.executionComplete(contextId)
-    )
   }
 
   it should "rename a project" in {
